Moves `ItemMetadata` to a "generic open class" for all VMs usage.
- Adds `CollectionCard` widget;
- Adds `buildHeroCollectionCardFromItem`, preparing for future animations using the `CollectionCard`;
- Adds `animations` file.
