+++ lib/application/constants/animations.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/animation.dart';
+
+const defaultAnimationCurve = Curves.easeInOutCubic;
+const defaultAnimatableProgressDuration = Duration(seconds: 1);
+
+const terminalFadeTransitionDuration = Duration(milliseconds: 300);
+const terminalActionsTransitionDuration = Duration(milliseconds: 300);
